Make CompactHistoryBlock size variable (v21.04.0)
When needed CompactHistoryBlock can now increase from 256kb to a bigger
size.

BUG: 436031

+++ src/history/compact/CompactHistoryLine.h
@@ -35,9 +35,9 @@ public:
 protected:
     CompactHistoryBlockList &_blockListRef;
     CharacterFormat *_formatArray;
-    quint16 _length;
+    int _length;
     uint    *_text;
-    quint16 _formatLength;
+    int _formatLength;
     bool _wrapped;
 };
